refactor: report 도메인 dto 리팩토링
