+++ src/main/java/coffeemeet/server/admin/presentation/dto/GroupReportHTTP.java
@@ -1,0 +1,19 @@
+package coffeemeet.server.admin.presentation.dto;
+
+import static lombok.AccessLevel.PRIVATE;
+
+import coffeemeet.server.report.service.dto.GroupReportDto;
+import java.util.List;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = PRIVATE)
+public final class GroupReportHTTP {
+
+  public record Response(List<GroupReportDto> groupReports) {
+
+    public static Response from(List<GroupReportDto> responses) {
+      return new Response(responses);
+    }
+  }
+
+}
