+++ src/main/java/coffeemeet/server/report/service/dto/GroupReportDto.java
@@ -1,10 +1,12 @@
 package coffeemeet.server.report.service.dto;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
 
 public record GroupReportDto(
     String reporterNickname,
     Long reportId,
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss.SSS")
     LocalDateTime createdAt
 ) {
