+++ src/main/java/coffeemeet/server/report/service/dto/ReportDto.java
@@ -2,6 +2,7 @@ package coffeemeet.server.report.service.dto;
 
 import coffeemeet.server.chatting.current.domain.ChattingRoom;
 import coffeemeet.server.user.domain.User;
+import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
 
 public record ReportDto(
@@ -9,6 +10,7 @@ public record ReportDto(
     String chattingRoomName,
     Long targetedId,
     Long chattingRoomId,
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss.SSS")
     LocalDateTime createdAt
 ) {
